Moving sourceFile object in his own file
