+++ internal/arduino/builder/internal/detector/detector.go
@@ -502,97 +502,6 @@ func findIncludeForOldCompilers(source string) string {
 	return ""
 }
 
-type sourceFile struct {
-	// Path to the source file within the sketch/library root folder
-	relativePath *paths.Path
-
-	// ExtraIncludePath contains an extra include path that must be
-	// used to compile this source file.
-	// This is mainly used for source files that comes from old-style libraries
-	// (Arduino IDE <1.5) requiring an extra include path to the "utility" folder.
-	extraIncludePath *paths.Path
-
-	// The source root for the given origin, where its source files
-	// can be found. Prepending this to SourceFile.RelativePath will give
-	// the full path to that source file.
-	sourceRoot *paths.Path
-
-	// The build root for the given origin, where build products will
-	// be placed. Any directories inside SourceFile.RelativePath will be
-	// appended here.
-	buildRoot *paths.Path
-}
-
-// Equals fixdoc
-func (f *sourceFile) Equals(g *sourceFile) bool {
-	return f.relativePath.EqualsTo(g.relativePath) &&
-		f.buildRoot.EqualsTo(g.buildRoot) &&
-		f.sourceRoot.EqualsTo(g.sourceRoot)
-}
-
-// makeSourceFile containing the given source file path within the
-// given origin. The given path can be absolute, or relative within the
-// origin's root source folder
-func makeSourceFile(
-	sourceDir *paths.Path,
-	buildDir *paths.Path,
-	sourceFilePath *paths.Path,
-	extraIncludePath ...*paths.Path,
-) (*sourceFile, error) {
-	res := &sourceFile{
-		buildRoot:  buildDir,
-		sourceRoot: sourceDir,
-	}
-
-	if len(extraIncludePath) > 1 {
-		panic("only one extra include path allowed")
-	}
-	if len(extraIncludePath) > 0 {
-		res.extraIncludePath = extraIncludePath[0]
-	}
-	// switch o := origin.(type) {
-	// case *sketch.Sketch:
-	// 	res.buildRoot = sketchBuildPath
-	// 	res.sourceRoot = sketchBuildPath
-	// case *libraries.Library:
-	// 	res.buildRoot = librariesBuildPath.Join(o.DirName)
-	// 	res.sourceRoot = o.SourceDir
-	// 	res.extraIncludePath = o.UtilityDir
-	// default:
-	// 	panic("Unexpected origin for SourceFile: " + fmt.Sprint(origin))
-	// }
-
-	if sourceFilePath.IsAbs() {
-		var err error
-		sourceFilePath, err = res.sourceRoot.RelTo(sourceFilePath)
-		if err != nil {
-			return nil, err
-		}
-	}
-	res.relativePath = sourceFilePath
-	return res, nil
-}
-
-// ExtraIncludePath fixdoc
-func (f *sourceFile) ExtraIncludePath() *paths.Path {
-	return f.extraIncludePath
-}
-
-// SourcePath fixdoc
-func (f *sourceFile) SourcePath() *paths.Path {
-	return f.sourceRoot.JoinPath(f.relativePath)
-}
-
-// ObjectPath fixdoc
-func (f *sourceFile) ObjectPath() *paths.Path {
-	return f.buildRoot.Join(f.relativePath.String() + ".o")
-}
-
-// DepfilePath fixdoc
-func (f *sourceFile) DepfilePath() *paths.Path {
-	return f.buildRoot.Join(f.relativePath.String() + ".d")
-}
-
 // LibrariesLoader todo
 func LibrariesLoader(
 	useCachedLibrariesResolution bool,
